+++ src/bsp/generic-vxworks/src/bsp_console.c
@@ -40,7 +40,7 @@
    OS_BSP_ConsoleOutput_Impl
    See full description in header
  ------------------------------------------------------------------*/
-void OS_BSP_ConsoleOutput_Impl(const char *Str, uint32 DataLen)
+void OS_BSP_ConsoleOutput_Impl(const char *Str, size_t DataLen)
 {
     while (DataLen > 0)
     {
@@ -58,5 +58,3 @@ void OS_BSP_ConsoleSetMode_Impl(uint32 ModeBits)
 {
     /* ignored; not implemented */
 }
-
-
