+++ src/bsp/pc-rtems/src/bsp_console.c
@@ -41,7 +41,7 @@
    OS_BSP_ConsoleOutput_Impl
    See full description in header
  ------------------------------------------------------------------*/
-void OS_BSP_ConsoleOutput_Impl(const char *Str, uint32 DataLen)
+void OS_BSP_ConsoleOutput_Impl(const char *Str, size_t DataLen)
 {
     /* writes the raw data directly to STDOUT_FILENO (unbuffered) */
     write(STDOUT_FILENO, Str, DataLen);
@@ -55,5 +55,3 @@ void OS_BSP_ConsoleSetMode_Impl(uint32 ModeBits)
 {
     /* no-op on RTEMS */
 }
-
-
