+++ src/os/portable/os-impl-no-network.c
@@ -28,14 +28,11 @@
  *
  */
 
-#include <osapi.h>
 #include "os-shared-network.h"
-
 
 /****************************************************************************************
                                     Network API
  ***************************************************************************************/
-
 
 /*----------------------------------------------------------------
  *
@@ -45,13 +42,10 @@
  *           See prototype for argument/return detail
  *
  *-----------------------------------------------------------------*/
-int32 OS_NetworkGetID_Impl             (int32 *IdBuf)
+int32 OS_NetworkGetID_Impl(int32 *IdBuf)
 {
     return OS_ERR_NOT_IMPLEMENTED;
 } /* end OS_NetworkGetID_Impl */
-
-
-
 
 /*----------------------------------------------------------------
  *
@@ -61,10 +55,7 @@ int32 OS_NetworkGetID_Impl             (int32 *IdBuf)
  *           See prototype for argument/return detail
  *
  *-----------------------------------------------------------------*/
-int32 OS_NetworkGetHostName_Impl       (char *host_name, uint32 name_len)
+int32 OS_NetworkGetHostName_Impl(char *host_name, size_t name_len)
 {
     return OS_ERR_NOT_IMPLEMENTED;
 } /* end OS_NetworkGetHostName_Impl */
-
-
-
